Fp10 104 day 3

+++ UnitFiveProject/src/main/java/ENTITY/LoggedInUserId.java
@@ -1,5 +1,5 @@
 package ENTITY;
 
 public class LoggedInUserId {
-	static public long loggedInUserId;
+	static public int loggedInUserId;
 }
